# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Condo Parking Spot Sharing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  apartment: string;
  phone: string;
  carPlate?: string | null;
  wantsToRequestSpot: boolean;
  hasParkingSpot: boolean;
  parkingSpotNumber?: string | null;
  phoneVerified: boolean;
  createdAt: string;
}

export interface RegisterUserInput {
  name: string;
  apartment: string;
  phone: string;
  carPlate?: string | null;
  wantsToRequestSpot?: boolean;
  hasParkingSpot?: boolean;
  parkingSpotNumber?: string | null;
}

export interface LoginInput {
  phone: string;
}

export interface UpdateProfileInput {
  carPlate?: string | null;
  wantsToRequestSpot?: boolean;
  hasParkingSpot?: boolean;
  parkingSpotNumber?: string | null;
}

export type ParkingSpotSpotType =
  (typeof ParkingSpotSpotType)[keyof typeof ParkingSpotSpotType];

export const ParkingSpotSpotType = {
  ONE_TIME: "ONE_TIME",
  RECURRING: "RECURRING",
} as const;

export type ParkingSpotStatus =
  (typeof ParkingSpotStatus)[keyof typeof ParkingSpotStatus];

export const ParkingSpotStatus = {
  AVAILABLE: "AVAILABLE",
  PENDING_CONFIRMATION: "PENDING_CONFIRMATION",
  OCCUPIED: "OCCUPIED",
  FINISHED: "FINISHED",
} as const;

export interface ParkingSpot {
  id: number;
  userId: number;
  userName: string;
  userApartment: string;
  userPhone: string;
  spotType: ParkingSpotSpotType;
  daysOfWeek?: string[] | null;
  availableFrom: string;
  availableUntil: string;
  date?: string | null;
  status: ParkingSpotStatus;
  interestedUserId?: number | null;
  interestedUserName?: string | null;
  interestedUserPhone?: string | null;
  interestedUserApartment?: string | null;
  approvalToken?: string | null;
  occupantName?: string | null;
  occupantApartment?: string | null;
  carPlate?: string | null;
  expectedExitTime?: string | null;
  createdAt: string;
}

export type CreateSpotInputSpotType =
  (typeof CreateSpotInputSpotType)[keyof typeof CreateSpotInputSpotType];

export const CreateSpotInputSpotType = {
  ONE_TIME: "ONE_TIME",
  RECURRING: "RECURRING",
} as const;

export interface CreateSpotInput {
  userId: number;
  spotType?: CreateSpotInputSpotType;
  daysOfWeek?: string[] | null;
  availableFrom: string;
  availableUntil: string;
  date?: string | null;
}

export interface ExpressInterestInput {
  interestedUserId: number;
}

export interface ConfirmOccupationInput {
  occupantName: string;
  occupantApartment: string;
  carPlate: string;
  expectedExitTime: string;
}

export type SpotRequestStatus =
  (typeof SpotRequestStatus)[keyof typeof SpotRequestStatus];

export const SpotRequestStatus = {
  open: "open",
  matched: "matched",
} as const;

export interface SpotRequest {
  id: number;
  userId: number;
  userName: string;
  userApartment: string;
  userPhone: string;
  date: string;
  startTime: string;
  endTime: string;
  reason?: string | null;
  status: SpotRequestStatus;
  offeredByUserId?: number | null;
  offeredByUserName?: string | null;
  offeredByUserPhone?: string | null;
  offeredByUserApartment?: string | null;
  createdAt: string;
}

export interface CreateSpotRequestInput {
  userId: number;
  date: string;
  startTime: string;
  endTime: string;
  reason?: string | null;
}

export interface OfferSpotInput {
  offeredByUserId: number;
}

export type ApproveBookingParams = {
  spotId: number;
  token: string;
};