# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Condo Parking Spot Sharing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /spots:
    get:
      operationId: getAvailableSpots
      tags: [spots]
      summary: Get all parking spots for today (one-time and recurring)
      responses:
        "200":
          description: List of spots today
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingSpot"

    post:
      operationId: createSpot
      tags: [spots]
      summary: Share a parking spot (one-time or recurring)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpotInput"
      responses:
        "201":
          description: Spot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
        "400":
          description: Phone not verified or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already has a spot listed today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /spots/approve:
    get:
      operationId: approveBooking
      tags: [spots]
      summary: Approve a booking via token link (owner clicks WhatsApp link)
      parameters:
        - name: spotId
          in: query
          required: true
          schema:
            type: integer
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking approved, spot now OCCUPIED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Spot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /spots/{id}:
    delete:
      operationId: removeSpot
      tags: [spots]
      summary: Remove a parking spot listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spot removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Spot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /spots/{id}/interest:
    post:
      operationId: expressInterest
      tags: [spots]
      summary: Express interest in a spot (AVAILABLE -> PENDING_CONFIRMATION)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpressInterestInput"
      responses:
        "200":
          description: Status updated to PENDING_CONFIRMATION with approval token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
        "400":
          description: User must verify phone or add car plate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Spot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Spot not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /spots/{id}/confirm:
    post:
      operationId: confirmOccupation
      tags: [spots]
      summary: Confirm spot occupation (PENDING_CONFIRMATION -> OCCUPIED)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmOccupationInput"
      responses:
        "200":
          description: Status updated to OCCUPIED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
        "404":
          description: Spot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Spot not in PENDING_CONFIRMATION state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /spots/{id}/vacate:
    post:
      operationId: vacateSpot
      tags: [spots]
      summary: Mark spot as vacated (OCCUPIED -> FINISHED or AVAILABLE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status updated to FINISHED or AVAILABLE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingSpot"
        "404":
          description: Spot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Spot not in OCCUPIED state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ParkingSpot:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userApartment:
          type: string
        userPhone:
          type: string
        spotType:
          type: string
          enum: [ONE_TIME, RECURRING]
        daysOfWeek:
          type: array
          nullable: true
          items:
            type: string
        availableFrom:
          type: string
        availableUntil:
          type: string
        date:
          type: string
          nullable: true
        status:
          type: string
          enum: [AVAILABLE, PENDING_CONFIRMATION, OCCUPIED, FINISHED]
        interestedUserId:
          type: integer
          nullable: true
        interestedUserName:
          type: string
          nullable: true
        interestedUserPhone:
          type: string
          nullable: true
        interestedUserApartment:
          type: string
          nullable: true
        approvalToken:
          type: string
          nullable: true
        occupantName:
          type: string
          nullable: true
        occupantApartment:
          type: string
          nullable: true
        carPlate:
          type: string
          nullable: true
        expectedExitTime:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - userApartment
        - userPhone
        - spotType
        - availableFrom
        - availableUntil
        - status
        - createdAt
    CreateSpotInput:
      type: object
      properties:
        userId:
          type: integer
        spotType:
          type: string
          enum: [ONE_TIME, RECURRING]
        daysOfWeek:
          type: array
          nullable: true
          items:
            type: string
        availableFrom:
          type: string
        availableUntil:
          type: string
        date:
          type: string
          nullable: true
      required: [userId, availableFrom, availableUntil]
    ExpressInterestInput:
      type: object
      properties:
        interestedUserId:
          type: integer
      required: [interestedUserId]
    ConfirmOccupationInput:
      type: object
      properties:
        occupantName:
          type: string
        occupantApartment:
          type: string
        carPlate:
          type: string
        expectedExitTime:
          type: string
      required: [occupantName, occupantApartment, carPlate, expectedExitTime]
...